chore: Move Sort class
This moves the Sort class and the interface into the Core directory.
