+++ tests/Core/SortTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types = 1);
 
-namespace Hypefactors\ElasticBuilder\Tests\Sort;
+namespace Hypefactors\ElasticBuilder\Tests\Core;
 
+use Hypefactors\ElasticBuilder\Core\Sort;
 use Hypefactors\ElasticBuilder\Script\Script;
-use Hypefactors\ElasticBuilder\Sort\Sort;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use stdClass;
